Explicitly mark objects of structured type with base class data
Base class data is vended as "just another child", which is mostly OK, except sometimes it's useful to know when the base class is being parsed vs. a member variable
Relying on the name "super" isn't actually correct, but at the Mirror level we actually know base class vs. ivar

This change propagates that knowledge via adding an extra byte to the payload for structured data in the logger format, such that the logger can tell consumers whether the 0th child is the base class or the first member variable

rdar://problem/19072746