Began implementing log data encoding.
Created a `LogEncoder` class which is capable of encoding the core log data types.
Created an unimplemented `LogEntry.encode(with:format:)` method which will soon encode a `LogEntry` into a given `LogEncoder` with the given `Format`.
Implemented the `LogPacket.encode(inFormat:)` function to properly encode its header and delegate to the `LogEntry.encode(with:format:)`.