Implemented the initial version of `LogEntry(describing:name:)`.
Implemented support for the following cases:

  - Types which implement `CustomPlaygroundRepresentable`
  - Types which implement `CustomPlaygroundQuickLookable`/`_DefaultCustomPlaygroundQuickLookable`
  - Fallback structural view for other types

There's also an unimplemented placeholder for types which implement the Objective-C method `-debugQuickLookObject`.
Aside from that, this is also an incomplete implementation as it does not handle errors very well -- for example, it just calls `fatalError` if converting from a `PlaygroundQuickLook` to a `LogEntry.OpaqueRepresentation` is not possible when it should instead throw an error.