Began replacing `LogEntry.OpaqueRepresentation` with a protocol.
Introduced the `OpaqueLogEntryRepresentation` protocol, which simply requires an `encode(into:usingFormat:)` method.
Made the `opaque` case of `LogEntry` take an `OpaqueLogEntryRepresentation` existential instead of an `OpaqueRepresentation` value.
Deprecated `LogEntry.OpaqueRepresentation`, but in lieu of transitioning anything off of it in this commit, simply made it conform to `OpaqueLogEntryRepresentation` and updated a couple sites accordingly.

Follow-on commits will introduce new types which conform to `OpaqueLogEntryRepresentation` and therefore remove `LogEntry.OpaqueRepresentation`.