Fully replaced the old `OpaqueRepresentation` enum with the `OpaqueLogEntryRepresentation` protocol (now typealiased as `LogEntry.OpaqueRepresentation`).
Implemented the new `OpaqueRepresentation` protocol on types which represent the current-supported opaque log entries. (With a couple notable exceptions: while NS/UIView and NS/UIImage purport to be supported, support for rendering images into log entries is not yet implemented and just calls `fatalError`.)
This is a more flexible approach which will make it easier to add support for other core types (for instance, support for logging CoreImage types directly).
Since it was no longer needed, deleted the old `OpaqueRepresentation` enum and its encoding support.