Replace `initializePlaygroundLogger()` with a modern initialization function.
This new function, `initializePlaygroundLogger(clientVersion:sendData:)`, has clients pass in a version and a function which will be used to send data to the host.
Instead of being declared with `@_silgen_name("playground_logger_initialize")`, this new function is declared with `@_cdecl("PGLInitializePlaygroundLogger")` as it is C-compatible.
A legacy entrypoint with `@_silgen_name("playground_logger_initialize")` will be added in a follow-up commit.