Implemented "legacy" entrypoints for PlaygroundLogger.
These functions match those used by LLDB in its playground preamble, allowing this framework to be used as a drop-in replacement for the legacy PlaygroundLogger framework.
Since these functions return NSData objects, they do not call the `sendData` function. As a result, the legacy initialization function sets `sendData` to a stub which calls through to `fatalError`.