Ported the tests from the legacy PlaygroundLogger implementation to the new one.
This includes pulling in the legacy PlaygroundLogger's decoder, and any dependencies it had from PlaygroundLogger. (This is isolated to a single file in the test suite.)
Most of the tests pass (thanks to a few fixes that landed in the last few commits), but there are a couple failures which will need to be addressed.