[PlaygroundSupport] De-duplicated much of the test files between macOS, iOS, and tvOS, and fixed the build for tests.
PlaygroundSupport and XCPlayground each had a lot of duplication in their test suites, when all that was different was UIView vs. NSView.
This makes each test suite use a single file for the live view-related tests, and also renames the test files to be more functionality-based.

This is in preparation to migrate the PlaygroundSupport project to Swift 4.2.