[PlaygroundLogger] Implemented a check to avoid trying to log nil-containing non-Optional objects.
It is undefined behavior for a non-Optional object to be nil in Swift, and at present, PlaygroundLogger triggers crashes in the runtime when trying to log such objects.
The most common way for such an object to exist is for an Objective-C API to be annotated as returning nonnull but in fact returns nil.
In a regular Swift program, this can be worked around by explicitly assigning the value to an Optional.
PlaygroundLogger cannot do that as it only receives an Any, and assigning to Any? does not recover the nil.

As a result, inspect the contents of the Any's existential container (i.e. the thing in memory that is the Any itself, not the value inside of the Any) to determine if it contains a non-Optional object whose value is nil.

This commit includes additional tests to ensure that we can log this broken value without crashing and that we can log a nil Optional object correctly.

This addresses <rdar://problem/56370098>.

(cherry picked from commit 1cad75496955bacbba92b813dbe4edf9c548d4b6)