[PlaygroundLogger] Added support for an environment variable which controls the maximum depth of logging.
Logging previously unconditionally used a maximum depth of 2 when using one of the regular entry points (because logging would use the default policy).
This commit allows clients to set `LOGGER_DEPTH` in the environment when launching the process performing logging; if this is set to a valid integer, it will override the default depth of 2.

This addresses <rdar://problem/67252287>.